refactor: 장기 / 단기 컴포넌트 분리
